add $DVDREAD_CFLAGS in "master" dvdread patch
Original patch by Rich Mattes <richmattes@gmail.com>

This was added to xbmc 12 in 4b47d8d256fbce87e73b21aa8f1ebd362b642937,
but I neglected to adjust the "nightly" patch in that commit. Do so
here.

(We need to get rid of the "nightly" patches in the tree, now that we
have an entire "nightly" branch in git. Sigh.)